Buttons are working correctly now
They stretch to the sides of the canvas, but are the correct height on both BeginGame and GameOver screens.  Quit didn't work because the code for running under the Editor was #if disabled.  The BeginGame didn't have an EventSystem on it, which is required to interact with buttons, but apparently not colliders.