Waves are now working again
I changed WaveConfig to inherit from MonoBehavior instead of ScriptableObject.  However, Start() was still not being called, so I changed it to be BuildShips() and that is called for each WaveConfig object from SpawnAllEnemies() in EnemySpawner.  This gets rid of the exception I was getting because the List of ships in the WaveConfig object was never populated.