Made changes to properly destroy objects now
I did see one exception where it says that when I was creating an enemy ship for a wave that the object had already been destroyed, but that would seem to be impossible.  A ship only gets destroyed if I have shot it, but I can't have shot it if it hasn't been created yet.  Catch-22.