Got delay working before going to GameOver scene
When player dies (takes two bombs to kill it) it is not going to GameOver scene, but it does if we collide with an ememy ship.

I also now make the name of all instantiated objects unique for bombs and enemies to better idetify them in the log.  I think we have an issue when a bomb is dropped that it kills the ship that dropped it.  Maybe need to adjust it so it is beyond the ship when it starts?