Background music added
Background music has been added and is played by a Singleton object in a loop.

Bugs that I noticed: Score is not carrying over to the Game Over scene.  If you click on Start Over in the Game Over scene you can't kill any of the enemy ships, although you can destroy the bombs.  You only see an explosion event for the bombs being destroyed.  Not sure if the player can be destroyed.