Score is now working!
I have the score working so you see it on the Game Over screen.  If you start playing again it goes to 0 and you are able to shoot enemies and bombs again.

One bug seems to be where a ship gets destroyed, but we are still spawning it and that causes us to access a destroyed object.

Another bug seems to be that I see enemy ships disappear when they haven't been hit by anything.  Not sure if that adds to my score or not.